#17: Add all metadata tables, refactoring
